protocol-native: add access checks
Call the hooks in various places to check if an action is allowed or
not.

The hook can return PA_HOOK_OK to allow an action, PA_HOOK_STOP to
deny an action.

Returning PA_HOOK_CANCEL will handle the result asynchronously.
This is implemented by saving the state of the current command and
exiting the command (going back to the mainloop) without reply.
When the access check completes, it will call async_finish_cb, which
will use the saved state to call the original command again. The access
check will eventually return OK or STOP and the command can complete
and send a reply.