module-access: add example access module
Add an example access module that only allows access to the objects
created by the owner client.

The code is structured in such a way that multiple profiles can be
made and that a profile can be activated based on the properties of
a client.

Events need special handling, we don't want to send events to clients
about objects they are not allowed to see. We need to be careful because
we can't inspect the owner of an object anymore in a _REMOVE event for
that object. We fix this by keeping a list of all the objects for which
we were allowed to notify a CHANGE or NEW event. We then only send
REMOVE events to objects from this list.