Add portal support
Call out to org.freedesktop.portal.Device.AccessDevice to
find if the user wants to let the app play or record sound.

Note that this currently does not work correctly, since pulseaudio
does not make the proper pid available that has been obtained
by peer-credentials. The pid that we get out of the proplist is
from inside the sandboxes pid namespace, and thus meaningless
on the outside.